+++ shelfsense/src/pages/StatisticsPage.jsx
@@ -4,7 +4,8 @@ import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterDateFns } from '@mui/x-date-pickers/AdapterDateFns';
 import { useEffect, useState } from "react";
 import { Sidebar } from "../components/sidebar/sidebar.jsx";
-import { calculateMonthlyStockUsage, calculateMonthlyEarnings } from "../util/services/SalesOrdersService";
+import { calculateMonthlyEarnings } from "../util/services/SalesOrders/calculateMonthlyEarnings.js";
+import { calculateMonthlyStockUsage } from "../util/services/SalesOrders/calculateMonthlyStockUsage.js";
 import { BarChart, LineChart } from '@mui/x-charts';
 import useComponentsStore from "../stores/useComponentsStore.js";
 import useProductsStore from "../stores/useProductsStore.js";
